requirements.txt
pandas>=2.0.0
numpy>=1.26.0
pyarrow>=15.0.0
requests>=2.31.0
ccxt>=4.3.98
python-dotenv>=1.0.1
APScheduler>=3.10.4
pytrends>=4.9.2
lightgbm>=4.3.0
scikit-learn>=1.4.2
torch>=2.1.0
scipy>=1.11.4
joblib>=1.3.2
